[ANDROSDK-1060] Rename configuration package to internal
