+++ core/src/main/java/org/hisp/dhis/android/core/configuration/internal/ConfigurationManagerImpl.java
@@ -26,7 +26,7 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.core.configuration;
+package org.hisp.dhis.android.core.configuration.internal;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
