+++ core/src/test/java/org/hisp/dhis/android/core/configuration/internal/ConfigurationManagerShould.java
@@ -26,7 +26,7 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.core.configuration;
+package org.hisp.dhis.android.core.configuration.internal;
 
 import org.hisp.dhis.android.core.arch.storage.internal.SecureStore;
 import org.junit.Before;
